+++ src/nova/trame/view/components/ornl/neutron_data_selector.py
@@ -201,6 +201,7 @@ class NeutronDataSelector(DataSelector):
         self._vm.directories_bind.connect(self._directories_name)
         self._vm.datafiles_bind.connect(self._datafiles_name)
         self._vm.reset_bind.connect(self.reset)
+        self._vm.reset_grid_bind.connect(self._reset_rv_grid)
 
         self._vm.update_view()
 
@@ -256,8 +257,12 @@ class NeutronDataSelector(DataSelector):
             @self.state.change(self._experiment[0].split(".")[0])
             def on_experiment_change(**kwargs: Any) -> None:
                 experiment = rgetdictvalue(kwargs, self._experiment[0])
-                if experiment != self._last_experiment:
+                if experiment and experiment != self._last_experiment:
                     self._last_experiment = experiment
+                    # See the note in the update_experiment method for why we call this twice.
+                    self._vm.set_binding_parameters(
+                        experiment=set_state_param(self.state, (self._selected_experiment_name,), ""),
+                    )
                     self._vm.set_binding_parameters(
                         experiment=set_state_param(self.state, (self._selected_experiment_name,), experiment)
                     )
@@ -304,6 +309,10 @@ class NeutronDataSelector(DataSelector):
         self._vm.reset()
 
     def update_experiment(self, experiment: str) -> None:
+        # Setting the experiment to an empty string forces the treeview to clear it's selection state.
+        self._vm.set_binding_parameters(
+            experiment=set_state_param(self.state, (self._selected_experiment_name,), ""),
+        )
         self._vm.set_binding_parameters(
             experiment=set_state_param(self.state, (self._selected_experiment_name,), experiment),
         )