+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 * Added data_source and projection parameters to NeutronDataSelector to allow populating data files from ONCat (thanks to Andrew Ayres and John Duggan).
 
+### nova-trame, 0.25.3
+
+* Clearing NeutronDataSelector file selections will no longer send null/None values to the state (thanks to John Duggan).
+
+### nova-trame, 0.25.2
+
+* NeutronDataSelector should now reset its state properly when changing the instrument or experiment (thanks to John Duggan).
+
 ### nova-trame, 0.25.1
 
 * ExecutionButtons now supports binding to stop_btn and download_btn parameters (thanks to John Duggan).